refactor: 이메일 프로퍼티 패키지 변경
